requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.3.0

# Interactive dashboards
ipywidgets>=7.6.0
jupyter>=1.0.0

# Note: LLM CLI tools (claude, gemini, openai) must be installed separately
# See README.md for installation instructions

# Optional: LLM API SDKs for direct API calls (--api-model mode)
# Install these to use API calls instead of CLI tools:
anthropic>=0.39.0
google-generativeai>=0.8.0
openai>=1.50.0
